+++ src/pages/popup/Sequence/pages/Multiple/CompoundMultipleTest.tsx
@@ -1,4 +1,4 @@
-import { APIWithParamsOrBody } from "@/pages/content/modules/getApiList2";
+import { APIWithParamsOrBody } from "@/pages/content/modules/getApiList";
 import { HandleCodeReturn } from "@/pages/popup/hooks/Request/useHandleCode";
 import Body from "@/pages/popup/ui/Body";
 import ModalCodeBlock from "@/pages/popup/ui/CodeBlockView";
