+++ src/pages/popup/shared/util/convertSelectedAPI.test.ts
@@ -1,4 +1,4 @@
-import { API } from "@/pages/content/modules/getApiList2";
+import { API } from "@/pages/content/modules/getApiList";
 import { SwaggerDocs } from "../api/docs";
 import { convertSelectedAPI } from "./convertSelectedAPI";
