+++ src/widgets/api-list/ui/normal-list/ApiList.tsx
@@ -1,4 +1,4 @@
-import { API, ApiList } from "@/pages/content/modules/getApiList2";
+import { API, ApiList } from "@/pages/content/modules/getApiList";
 import { vars } from "@/shared/ui/styles/theme.css";
 import APIItem from "../../../../pages/popup/ui/APIItem/APIItem";
 import { ApiListStyle } from "../apiList.css";
